Vizart Application/backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pillow==10.1.0
opencv-python==4.8.1.78
numpy==1.24.3
torch==2.1.1
torchvision==0.16.1
mediapipe==0.10.8
scikit-learn==1.3.2
matplotlib==3.7.2
rembg==2.0.50
tqdm==4.66.1
redis==5.0.1
celery==5.3.4
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1
boto3==1.34.0
python-dotenv==1.0.0
sentry-sdk==1.38.0
structlog==23.2.0